+++ database/migrations/2019_05_03_000002_create_subscriptions_table.php
@@ -19,6 +19,7 @@ class CreateSubscriptionsTable extends Migration
                 $table->unsignedBigInteger('user_id');
                 $table->string('name');
                 $table->string('stripe_id')->collation('utf8mb4_bin');
+                $table->string('stripe_status');
                 $table->string('stripe_plan');
                 $table->integer('quantity');
                 $table->timestamp('trial_ends_at')->nullable();