+++ tests/Integration/ChargesTest.php
@@ -2,8 +2,8 @@
 
 namespace Laravel\Cashier\Tests\Integration;
 
-use Stripe\Charge;
-use Stripe\Error\InvalidRequest;
+use Laravel\Cashier\Payment;
+use Laravel\Cashier\Exceptions\PaymentActionRequired;
 
 class ChargesTest extends IntegrationTestCase
 {
@@ -11,33 +11,22 @@ class ChargesTest extends IntegrationTestCase
     {
         $user = $this->createCustomer('customer_can_be_charged');
         $user->createAsStripeCustomer();
-        $user->updateCard('tok_visa');
 
-        $response = $user->charge(1000);
+        $response = $user->charge(1000, 'pm_card_visa');
 
-        $this->assertInstanceOf(Charge::class, $response);
-        $this->assertEquals(1000, $response->amount);
+        $this->assertInstanceOf(Payment::class, $response);
+        $this->assertEquals(1000, $response->rawAmount());
         $this->assertEquals($user->stripe_id, $response->customer);
-    }
-
-    public function test_customer_cannot_be_charged_with_custom_source()
-    {
-        $user = $this->createCustomer('customer_can_be_charged_with_custom_source');
-        $user->createAsStripeCustomer();
-
-        $this->expectException(InvalidRequest::class);
-
-        $user->charge(1000, ['source' => 'tok_visa']);
     }
 
     public function test_non_stripe_customer_can_be_charged()
     {
         $user = $this->createCustomer('non_stripe_customer_can_be_charged');
 
-        $response = $user->charge(1000, ['source' => 'tok_visa']);
+        $response = $user->charge(1000, 'pm_card_visa');
 
-        $this->assertInstanceOf(Charge::class, $response);
-        $this->assertEquals(1000, $response->amount);
+        $this->assertInstanceOf(Payment::class, $response);
+        $this->assertEquals(1000, $response->rawAmount());
         $this->assertNull($response->customer);
     }
 
@@ -45,7 +34,7 @@ class ChargesTest extends IntegrationTestCase
     {
         $user = $this->createCustomer('customer_can_be_charged_and_invoiced_immediately');
         $user->createAsStripeCustomer();
-        $user->updateCard('tok_visa');
+        $user->updateDefaultPaymentMethod('pm_card_visa');
 
         $user->invoiceFor('Laravel Cashier', 1000);
 
@@ -58,11 +47,29 @@ class ChargesTest extends IntegrationTestCase
     {
         $user = $this->createCustomer('customer_can_be_refunded');
         $user->createAsStripeCustomer();
-        $user->updateCard('tok_visa');
+        $user->updateDefaultPaymentMethod('pm_card_visa');
 
         $invoice = $user->invoiceFor('Laravel Cashier', 1000);
-        $refund = $user->refund($invoice->charge);
+        $refund = $user->refund($invoice->payment_intent);
 
         $this->assertEquals(1000, $refund->amount);
+    }
+
+    public function test_charging_may_require_an_extra_action()
+    {
+        $user = $this->createCustomer('charging_may_require_an_extra_action');
+        $user->createAsStripeCustomer();
+
+        try {
+            $user->charge(1000, 'pm_card_threeDSecure2Required');
+
+            $this->fail('Expected exception '.PaymentActionRequired::class.' was not thrown.');
+        } catch (PaymentActionRequired $e) {
+            // Assert that the payment needs an extra action.
+            $this->assertTrue($e->payment->requiresAction());
+
+            // Assert that the payment was for the correct amount.
+            $this->assertEquals(1000, $e->payment->rawAmount());
+        }
     }
 }