+++ tests/Integration/InvoicesTest.php
@@ -2,7 +2,7 @@
 
 namespace Laravel\Cashier\Tests\Integration;
 
-use Stripe\Invoice;
+use Laravel\Cashier\Invoice;
 use Laravel\Cashier\Exceptions\InvalidStripeCustomer;
 
 class InvoicesTest extends IntegrationTestCase
@@ -20,7 +20,7 @@ class InvoicesTest extends IntegrationTestCase
     {
         $user = $this->createCustomer('invoicing_fails_with_nothing_to_invoice');
         $user->createAsStripeCustomer();
-        $user->updateCard('tok_visa');
+        $user->updateDefaultPaymentMethod('pm_card_visa');
 
         $response = $user->invoice();
 
@@ -31,7 +31,7 @@ class InvoicesTest extends IntegrationTestCase
     {
         $user = $this->createCustomer('customer_can_be_invoiced');
         $user->createAsStripeCustomer();
-        $user->updateCard('tok_visa');
+        $user->updateDefaultPaymentMethod('pm_card_visa');
 
         $response = $user->invoiceFor('Laracon', 49900);