+++ tests/Integration/WebhooksTest.php
@@ -5,6 +5,9 @@ namespace Laravel\Cashier\Tests\Integration;
 use Stripe\Plan;
 use Stripe\Product;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Notification;
+use Laravel\Cashier\Notifications\ConfirmPayment;
+use Laravel\Cashier\Exceptions\PaymentActionRequired;
 
 class WebhooksTest extends IntegrationTestCase
 {
@@ -53,7 +56,7 @@ class WebhooksTest extends IntegrationTestCase
     public function test_cancelled_subscription_is_properly_reactivated()
     {
         $user = $this->createCustomer('cancelled_subscription_is_properly_reactivated');
-        $subscription = $user->newSubscription('main', static::$planId)->create('tok_visa');
+        $subscription = $user->newSubscription('main', static::$planId)->create('pm_card_visa');
         $subscription->cancel();
 
         $this->assertTrue($subscription->cancelled());
@@ -76,7 +79,7 @@ class WebhooksTest extends IntegrationTestCase
     public function test_subscription_is_marked_as_cancelled_when_deleted_in_stripe()
     {
         $user = $this->createCustomer('subscription_is_marked_as_cancelled_when_deleted_in_stripe');
-        $subscription = $user->newSubscription('main', static::$planId)->create('tok_visa');
+        $subscription = $user->newSubscription('main', static::$planId)->create('pm_card_visa');
 
         $this->assertFalse($subscription->cancelled());
 
@@ -92,5 +95,57 @@ class WebhooksTest extends IntegrationTestCase
         ])->assertOk();
 
         $this->assertTrue($subscription->fresh()->cancelled(), 'Subscription is still active.');
+    }
+
+    public function test_subscription_is_deleted_when_status_is_incomplete_expired()
+    {
+        $user = $this->createCustomer('subscription_is_deleted_when_status_is_incomplete_expired');
+        $subscription = $user->newSubscription('main', static::$planId)->create('pm_card_visa');
+
+        $this->assertCount(1, $user->subscriptions);
+
+        $this->postJson('stripe/webhook', [
+            'id' => 'foo',
+            'type' => 'customer.subscription.updated',
+            'data' => [
+                'object' => [
+                    'id' => $subscription->stripe_id,
+                    'customer' => $user->stripe_id,
+                    'status' => 'incomplete_expired',
+                ],
+            ],
+        ])->assertOk();
+
+        $this->assertEmpty($user->refresh()->subscriptions, 'Subscription was not deleted.');
+    }
+
+    public function test_payment_action_required_email_is_sent()
+    {
+        $user = $this->createCustomer('payment_action_required_email_is_sent');
+
+        try {
+            $user->newSubscription('main', static::$planId)->create('pm_card_threeDSecure2Required');
+
+            $this->fail('Expected exception '.PaymentActionRequired::class.' was not thrown.');
+        } catch (PaymentActionRequired $exception) {
+            Notification::fake();
+
+            $this->postJson('stripe/webhook', [
+                'id' => 'foo',
+                'type' => 'invoice.payment_action_required',
+                'data' => [
+                    'object' => [
+                        'id' => 'foo',
+                        'customer' => $user->stripe_id,
+                        'payment_intent' => $exception->payment->id,
+                    ],
+                ],
+            ])->assertOk();
+
+            Notification::assertSentTo($user, ConfirmPayment::class, function (ConfirmPayment $notification) use ($exception) {
+                return $notification->paymentId === $exception->payment->id &&
+                    $notification->amount === $exception->payment->amount();
+            });
+        }
     }
 }