+++ CHANGELOG.md
@@ -1,9 +1,48 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/cashier/compare/v9.3.3...master)
+## [Unreleased](https://github.com/laravel/cashier/compare/v10.0.0-beta.2...master)
+
+
+## [v10.0.0-beta.2](https://github.com/laravel/cashier/compare/v10.0.0-beta...v10.0.0-beta.2)
 
 ### Added
-- Allow Carbon 2 ([#607](https://github.com/laravel/cashier/pull/607))
+- Add latestPayment method on Subscription ([#705](https://github.com/laravel/cashier/pull/705))
+- Allow custom filename for invoice download ([#723](https://github.com/laravel/cashier/pull/723))
+
+### Changed
+- Improve stripe statuses ([#707](https://github.com/laravel/cashier/pull/707))
+- Refactor active subscription state ([#712](https://github.com/laravel/cashier/pull/712))
+- Return invoice object when applicable ([#711](https://github.com/laravel/cashier/pull/711))
+- Refactor webhook responses ([#722](https://github.com/laravel/cashier/pull/722))
+- Refactor confirm payment mail to notification ([#727](https://github.com/laravel/cashier/pull/727))
+
+### Fixed
+- Fix createSetupIntent ([#704](https://github.com/laravel/cashier/pull/704))
+- Fix subscription invoicing ([#710](https://github.com/laravel/cashier/pull/710))
+- Fix `null` return for `latestPayment` method ([#730](https://github.com/laravel/cashier/pull/730))
+
+### Removed
+- Remove unused `$customer` parameter on `updateQuantity` method ([#729](https://github.com/laravel/cashier/pull/729))
+
+
+## [v10.0.0-beta](https://github.com/laravel/cashier/compare/v9.3.5...v10.0.0-beta)
+
+Cashier 10.0 is a major release. Please review [the upgrade guide](UPGRADE.md) thoroughly.
+
+
+## [v9.3.5 (2019-07-30)](https://github.com/laravel/cashier/compare/v9.3.4...v9.3.5)
+
+### Changed
+- Remove old 5.9 version constraints ([c7664fc](https://github.com/laravel/cashier/commit/c7664fc90d0310d6fa3a52bec45e94868bff995d))
+
+### Fixed
+- Don't try and find a user when stripeId is null ([#721](https://github.com/laravel/cashier/pull/721))
+
+
+## [v9.3.4 (2019-07-29)](https://github.com/laravel/cashier/compare/v9.3.3...v9.3.4)
+
+### Changed
+- Updated version constraints for Laravel 6.0 ([4a4c5c2](https://github.com/laravel/cashier/commit/4a4c5c226bb98aa0726f57bb5970115d3eaab377))
 
 
 ## [v9.3.3 (2019-06-14)](https://github.com/laravel/cashier/compare/v9.3.2...v9.3.3)